+++ ledfx/api/playlists.py
@@ -149,32 +149,3 @@ class PlaylistsEndpoint(RestEndpoint):
             )
         return await self.invalid_request("Playlist not found")
 
-
-class PlaylistEndpoint(RestEndpoint):
-    """REST endpoint for single playlist retrieval."""
-
-    ENDPOINT_PATH = "/api/playlists/{id}"
-
-    async def get(self, id) -> web.Response:
-        if not hasattr(self._ledfx, "playlists"):
-            self._ledfx.playlists = PlaylistManager(self._ledfx)
-        p = self._ledfx.playlists.get_playlist(id)
-        if not p:
-            return await self.invalid_request("Playlist not found")
-
-        return await self.request_success(data={"playlist": p})
-
-
-class ActivePlaylistEndpoint(RestEndpoint):
-    """REST endpoint for reporting runtime state of playlists."""
-
-    ENDPOINT_PATH = "/api/playlists/active"
-
-    async def get(self) -> web.Response:
-        if not hasattr(self._ledfx, "playlists"):
-            self._ledfx.playlists = PlaylistManager(self._ledfx)
-        # If there is no active playlist, return state: null to match docs
-        if not self._ledfx.playlists._active_playlist_id:
-            return await self.request_success(data={"state": None})
-        state = await self._ledfx.playlists.get_state()
-        return await self.request_success(data={"state": state})
